Fix for mobs (rarely) walking through walls.
When a monster's move will cross a wall, the point on the wall for the
lines move_start->move_end and the wall line itself is calculated.
Currently the X,Y values for this point are often incorrect by < 0.1,
and if the end point is across a wall from the start point and the end
point is very close to the wall, this will allow the monster to
incorrectly move through it.

To fix this, the calculation is now done using doubles instead of floats
(fixing cases where c1 is large). Some cases remained after this and to
handle those, the ISINBOX macro now subtracts 0.1f from the min box
value and adds 0.1f to the max box value to ensure the point is
contained within.

+++ blakserv/roofile.c
@@ -462,25 +462,25 @@ bool BSPCanMoveInRoomTree(BspNode* Node, V2* S, V2* E, Wall** BlockWall)
           (distS < 0.0f && distE > 0.0f))
       {
          // get 2d line equation coefficients for infinite line through S and E
-         float a1, b1, c1;
+         double a1, b1, c1;
          a1 = E->Y - S->Y;
          b1 = S->X - E->X;
          c1 = a1 * S->X + b1 * S->Y;
 
          // get 2d line equation coefficients of splitter
-         float a2, b2, c2;
+         double a2, b2, c2;
          a2 = -Node->u.internal.A;
          b2 = -Node->u.internal.B;
          c2 = Node->u.internal.C;
 
-         float det = a1*b2 - a2*b1;
+         double det = a1*b2 - a2*b1;
 
          // shouldn't be zero at all, because distS and distE have different sign
          if (!ISZERO(det))
          {
-            // intersection point of infinite lines				
-            q.X = (b2*c1 - b1*c2) / det;
-            q.Y = (a1*c2 - a2*c1) / det;
+            // intersection point of infinite lines
+            q.X = (float)((b2*c1 - b1*c2) / det);
+            q.Y = (float)((a1*c2 - a2*c1) / det);
 
             // must be in boundingbox of SE
             if (ISINBOX(S, E, &q))
